Add files via upload
It's a mobile number calculator which adds the even and odd digits of your phone number. If the sum of the even digits is greater than the odd digits Access will be granted and vice versa. If the sum of odd and even digits is the same then it will generate a random integer which will be then checked if it's odd or even, if even then Access will be granted and if odd access will be denied.